some more checks for int overflows
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1857609 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/poifs/filesystem/POIFSDocument.java
@@ -154,7 +154,7 @@ public final class POIFSDocument implements POIFSViewable, Iterable<ByteBuffer> 
            }
        }
 
-       return (int)length;
+       return Math.toIntExact(length);
    }
    
    /**
